split buildingplan into a project
