+++ plugins/CMakeLists.txt
@@ -89,7 +89,7 @@ dfhack_plugin(autonestbox autonestbox.cpp LINK_LIBRARIES lua)
 dfhack_plugin(blueprint blueprint.cpp LINK_LIBRARIES lua)
 #dfhack_plugin(burrows burrows.cpp LINK_LIBRARIES lua)
 #dfhack_plugin(building-hacks building-hacks.cpp LINK_LIBRARIES lua)
-dfhack_plugin(buildingplan buildingplan.cpp LINK_LIBRARIES lua)
+add_subdirectory(buildingplan)
 #dfhack_plugin(changeitem changeitem.cpp)
 dfhack_plugin(changelayer changelayer.cpp)
 dfhack_plugin(changevein changevein.cpp)
