+++ plugins/buildingplan/CMakeLists.txt
@@ -2,10 +2,12 @@ project(buildingplan)
 
 set(COMMON_HDRS
     buildingplan.h
-    buildingplan-planner.h
-    buildingplan-rooms.h
+    itemfilter.h
+    plannedbuilding.h
 )
 set_source_files_properties(${COMMON_HDRS} PROPERTIES HEADER_FILE_ONLY TRUE)
 
-dfhack_plugin(buildingplan buildingplan.cpp buildingplan-planner.cpp
-    buildingplan-rooms.cpp ${COMMON_HDRS} LINK_LIBRARIES lua)
+dfhack_plugin(buildingplan
+    buildingplan.cpp buildingplan_cycle.cpp itemfilter.cpp plannedbuilding.cpp
+    ${COMMON_HDRS}
+    LINK_LIBRARIES lua)
